stub operation template
