+++ app/src/main/java/com/example/ignition/financetracker/entities/Operation.kt
@@ -14,10 +14,11 @@ data class Operation(@PrimaryKey(autoGenerate = true)
                      val date: Long,
                      val rate: BigDecimal = BigDecimal.ONE)
 
-data class OperationFilter(val from: Long,
-                           val to: Long,
-                           val income: Boolean,
-                           val expense: Boolean,
-                           val all: Boolean,
-                           val category: String,
-                           val anyCategory: Boolean)
+@Entity(tableName = "operation_template")
+data class OperationTemplate(@PrimaryKey(autoGenerate = true)
+                             val id: Long = 0,
+                             val walletName: String,
+                             val operationType: String,
+                             val sum: BigDecimal,
+                             val currency: String,
+                             val rate: BigDecimal = BigDecimal.ONE)
