+++ app/src/main/java/com/example/ignition/financetracker/ui/main/models.kt
@@ -15,5 +15,10 @@ data class WalletModel(val walletName: String,
                        val outcomeValue: BigDecimal,
                        val operations: List<Operation>)
 
-data class WalletOperationModel(val operation: Operation,
-                                val mainToSecondaryRate: BigDecimal)
+data class OperationFilter(val from: Long,
+                           val to: Long,
+                           val income: Boolean,
+                           val expense: Boolean,
+                           val all: Boolean,
+                           val category: String,
+                           val anyCategory: Boolean)
