+++ app/src/main/res/layout/dialog_addoperation.xml
@@ -90,6 +90,13 @@
             android:textSize="20sp" />
 
         <Switch
+            android:id="@+id/switch_saveAsTemplate"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:checked="false"
+            android:text="Сохранить как шаблон" />
+
+        <Switch
             android:id="@+id/switch_transactionPeriodDate"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
